Rename package com.microsoft.springframework.samples to com.azure.spring.samples, the track2 version can be used for better identification.
