+++ src/main/java/com/azure/spring/samples/dao/TodoItemRepository.java
@@ -1,0 +1,14 @@
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See LICENSE in the project root for
+ * license information.
+ */
+package com.azure.spring.samples.dao;
+
+import com.azure.spring.samples.model.TodoItem;
+import com.microsoft.azure.spring.data.cosmosdb.repository.DocumentDbRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TodoItemRepository extends DocumentDbRepository<TodoItem, String> {
+}
