+++ src/test/java/com/azure/spring/samples/model/TodoItemTest.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See LICENSE in the project root for
  * license information.
  */
-package com.microsoft.springframework.samples.model;
+package com.azure.spring.samples.model;
 
 import static org.junit.Assert.*;
